pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doaj"
version = "0.1.0"
description = "DOAJ project"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
dependencies = [
  "fastapi>=0.110",
  "polars>=0.20",
  "python-dotenv>=1.0",
  "requests>=2.31",
  "uvicorn>=0.25"
]
authors = [
  { name = "Ikhwan Arief", email = "ikhwanarief@gmail.com" }
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.4"
]

[project.scripts]
doaj = "doaj.main:main"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
addopts = "-q"
